@@ -234,22 +239,25 @@
     "localStorage": "1.0.3",
     "lodash": "4.17.2",
     "minimist": "1.2.0",
-    "poker-helper": "0.3.19",
+    "poker-helper": "0.3.20",
+    "prop-types": "^15.5.10",
     "pusher-js": "^4.0.0",
     "qrcode.react": "^0.6.1",
     "raven-js": "^3.13.1",
     "rc-slider": "^8.1.2",
-    "react": "15.4.2",
-    "react-dom": "15.4.2",
+    "react": "15.6.1",
+    "react-dom": "15.6.1",
+    "react-ga": "^2.2.0",
     "react-google-recaptcha": "^0.5.4",
     "react-helmet": "3.2.2",
     "react-intl": "2.1.5",
+    "react-onclickoutside": "^6.1.1",
     "react-rangeslider": "^2.1.0",
     "react-redux": "4.4.6",
     "react-router": "3.0.0",
     "react-router-redux": "4.0.6",
     "react-router-scroll": "0.4.1",
-    "react-test-renderer": "^15.4.2",
+    "react-test-renderer": "^15.6.1",
     "redux": "3.6.0",
     "redux-form": "^6.5.0",
     "redux-form-saga": "^0.1.0",
@@ -269,61 +277,72 @@
     "@storybook/addon-actions": "^3.1.2",
     "@storybook/addon-knobs": "^3.0.1",
     "@storybook/react": "^3.0.1",
-    "babel-cli": "6.18.0",
-    "babel-core": "6.21.0",
-    "babel-eslint": "7.1.1",
-    "babel-loader": "6.2.10",
-    "babel-plugin-dynamic-import-node": "1.0.0",
-    "babel-plugin-react-intl": "2.2.0",
+    "babel-cli": "6.24.1",
+    "babel-core": "6.24.1",
+    "babel-eslint": "7.2.3",
+    "babel-loader": "7.0.0",
+    "babel-plugin-dynamic-import-node": "1.0.2",
+    "babel-plugin-react-intl": "2.3.1",
     "babel-plugin-react-transform": "2.0.2",
-    "babel-plugin-transform-es2015-modules-commonjs": "6.18.0",
-    "babel-plugin-transform-react-constant-elements": "6.9.1",
-    "babel-plugin-transform-react-inline-elements": "6.8.0",
-    "babel-plugin-transform-react-remove-prop-types": "0.2.11",
-    "babel-preset-latest": "6.16.0",
-    "babel-preset-react": "6.16.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "6.24.1",
+    "babel-plugin-transform-react-constant-elements": "6.23.0",
+    "babel-plugin-transform-react-inline-elements": "6.22.0",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.5",
+    "babel-preset-env": "1.5.1",
+    "babel-preset-react": "6.24.1",
     "babel-preset-react-hmre": "1.1.1",
-    "babel-preset-stage-0": "6.16.0",
+    "babel-preset-stage-0": "6.24.1",
     "cheerio": "0.22.0",
-    "circular-dependency-plugin": "2.0.0",
-    "coveralls": "2.11.15",
-    "css-loader": "0.26.1",
-    "enzyme": "2.6.0",
-    "eslint": "3.11.1",
-    "eslint-config-airbnb": "13.0.0",
-    "eslint-config-airbnb-base": "10.0.1",
-    "eslint-import-resolver-webpack": "0.8.0",
-    "eslint-plugin-import": "2.2.0",
-    "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "6.7.1",
-    "eslint-plugin-redux-saga": "0.1.5",
+    "circular-dependency-plugin": "3.0.0",
+    "coveralls": "2.13.1",
+    "css-loader": "0.28.4",
+    "enzyme": "2.8.2",
+    "eslint": "3.19.0",
+    "eslint-config-airbnb": "15.0.1",
+    "eslint-config-airbnb-base": "11.2.0",
+    "eslint-import-resolver-webpack": "0.8.1",
+    "array-find": "^1.0.0",
+    "debug": "^2.6.8",
+    "enhanced-resolve": "~0.9.0",
+    "find-root": "^0.1.1",
+    "has": "^1.0.1",
+    "interpret": "^1.0.0",
+    "is-absolute": "^0.2.3",
+    "lodash.get": "^3.7.0",
+    "node-libs-browser": "^1.0.0",
+    "resolve": "^1.2.0",
+    "semver": "^5.3.0",
+    "eslint-plugin-import": "2.3.0",
+    "eslint-plugin-jsx-a11y": "5.0.3",
+    "eslint-plugin-react": "7.0.1",
+    "eslint-plugin-redux-saga": "0.3.0",
     "eventsource-polyfill": "0.9.6",
-    "exports-loader": "0.6.3",
-    "file-loader": "0.9.0",
-    "html-loader": "0.4.4",
-    "html-webpack-plugin": "2.24.1",
+    "exports-loader": "0.6.4",
+    "file-loader": "0.11.1",
+    "html-loader": "0.4.5",
+    "html-webpack-plugin": "2.28.0",
     "image-webpack-loader": "2.0.0",
-    "imports-loader": "0.6.5",
-    "jest-cli": "19.0.2",
-    "lint-staged": "3.2.1",
-    "ngrok": "2.2.4",
+    "imports-loader": "0.7.1",
+    "jest-cli": "20.0.4",
+    "lint-staged": "3.5.1",
+    "ngrok": "2.2.9",
     "nock": "^9.0.11",
-    "node-plop": "0.5.4",
+    "node-plop": "0.7.0",
     "null-loader": "0.1.1",
-    "offline-plugin": "4.5.2",
-    "plop": "1.7.3",
-    "pre-commit": "1.1.3",
-    "react-addons-shallow-compare": "^15.4.2",
-    "react-addons-test-utils": "15.4.2",
+    "offline-plugin": "4.8.1",
+    "plop": "1.8.0",
+    "pre-commit": "1.2.2",
+    "react-addons-shallow-compare": "^15.6.0",
+    "react-addons-test-utils": "15.6.0",
     "redux-saga-tester": "^1.0.186",
-    "rimraf": "2.5.4",
-    "shelljs": "0.7.5",
-    "sinon": "2.0.0-pre",
+    "rimraf": "2.6.1",
+    "shelljs": "0.7.7",
+    "sinon": "2.3.2",
     "storybook-addon-specifications": "^2.0.0",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "webpack": "2.2.0-rc.3",
-    "webpack-dev-middleware": "1.9.0",
-    "webpack-hot-middleware": "2.15.0"
+    "style-loader": "0.18.1",
+    "url-loader": "0.5.8",
+    "webpack": "2.6.1",
+    "webpack-dev-middleware": "1.10.2",
+    "webpack-hot-middleware": "2.18.0"
   }
 }
