@@ -1,37 +1,59 @@
 import React from 'react';
+import PropTypes from 'prop-types';
+import Link from '../Link';
 
 import {
  ItemWrapper,
  ItemIcon,
  ItemTitle,
+ LinkWrapper,
 } from './styles';
 
-const MenuItems = (props) => {
-  const { item } = props;
-  // if the menu is open, close it
+const MenuItem = ({ item, ...props }) => {
+  const children = [
+    <ItemIcon className={item.icon} aria-hidden key="0" />,
+    <ItemTitle key="1">
+      {item.title}
+    </ItemTitle>,
+  ];
+
   const handleClick = () => {
+    // if the menu is open, close it
     if (props.open) {
       props.toggleMenuOpen();
     }
-    item.onClick();
+    if (item.onClick) {
+      item.onClick();
+    }
   };
+
+  if (item.onClick) {
+    return (
+      <ItemWrapper
+        name={item.name}
+        disabled={item.disabled}
+        onClick={handleClick}
+      >
+        {children}
+      </ItemWrapper>
+    );
+  }
+
   return (
-    <ItemWrapper
+    <Link
+      to={item.to}
       name={item.name}
-      disabled={item.disabled}
       onClick={handleClick}
+      component={LinkWrapper}
     >
-      <ItemIcon className={item.icon} aria-hidden />
-      <ItemTitle>
-        {item.title}
-      </ItemTitle>
-    </ItemWrapper>
+      {children}
+    </Link>
   );
 };
-MenuItems.propTypes = {
-  item: React.PropTypes.object,
-  open: React.PropTypes.bool,
-  toggleMenuOpen: React.PropTypes.func,
+MenuItem.propTypes = {
+  item: PropTypes.object,
+  open: PropTypes.bool,
+  toggleMenuOpen: PropTypes.func,
 };
 
-export default MenuItems;
+export default MenuItem;
