@@ -7,71 +7,57 @@ import {
   number,
 } from '@storybook/addon-knobs';
 
-import tests, { combine } from '../tests';
-import {
-  atTable0,
-  atTable1,
-  atTable2,
-  amountToCall1,
-  amountToCall2,
-  amountToCall3,
-  minRaise0,
-  buttonBet0,
-  buttonBet1,
-  buttonRaise0,
-  buttonRaise1,
-} from '../tests2';
+import * as tests from '../tests';
 
 import ActionBar from '../../index';
 
-const stories = storiesOf('ActionBar', module);
-
 const createStory = (test) => {
-  stories.add(combine(test.describe, test.it), () => (
+  stories.add(tests.combine(test.describe, test.it), () => (
     <ActionBar {...test.props} />
   ));
 };
+
+const stories = storiesOf('ActionBar', module);
 
 stories.addDecorator(withKnobs);
 
 stories.add('Kitchen Sink', () => (
   <ActionBar
     active={boolean('active', true)}
-    amount={number('amount', 100)}
+    amount={number('amount', 200)}
     amountToCall={number('amountToCall', 100)}
+    buttonActive={text('buttonActive', '')}
     callAmount={number('callAmount', 0)}
     handleFold={() => {}}
     handleCall={() => {}}
     handleCheck={() => {}}
     handleBet={() => {}}
     isMyTurn={boolean('isMyTurn', true)}
-    minRaise={number('minRaise', 0)}
-    mode={text('mode', null)}
+    minRaise={number('minRaise', 200)}
+    mode={text('mode', '')}
     myStack={number('myStack', 10000)}
     params={{ tableAddr: '0x33' }}
+    potSize={number('potSize', 400)}
     setActionBarBetSlider={() => {}}
+    setActionBarButtonActive={() => {}}
+    setActionBarMode={() => {}}
     sliderOpen={boolean('sliderOpen', false)}
     state={text('state', 'flop')}
+    turnComplete={boolean('turnComplete', false)}
     updateAmount={() => {}}
     visible={boolean('visible', true)}
   />
 ));
 
-createStory(atTable0);
-createStory(atTable1);
-createStory(atTable2);
-createStory(amountToCall1);
-createStory(amountToCall2);
-createStory(amountToCall3);
-createStory(minRaise0);
-createStory(buttonBet0);
-createStory(buttonBet1);
-createStory(buttonRaise0);
-createStory(buttonRaise1);
-
-// iterate over tests and add a story for each
-tests.forEach((test) => {
-  stories.add(combine(test.describe, test.it), () => (
-    <ActionBar {...test.props} />
-  ));
-});
+createStory(tests.atTable0);
+createStory(tests.atTable1);
+createStory(tests.atTable2);
+createStory(tests.amountToCheck);
+createStory(tests.amountToCall0);
+createStory(tests.amountToCall1);
+createStory(tests.minRaise0);
+createStory(tests.buttonBet0);
+createStory(tests.buttonBet1);
+createStory(tests.buttonRaise0);
+createStory(tests.buttonRaise1);
+createStory(tests.actionDispatchRaise0);
