@@ -8,10 +8,9 @@ import { mount } from 'enzyme';
 import Button from '../index';
 
 const handleRoute = () => {};
-const href = 'http://mxstbr.com';
 const children = (<h1>Test</h1>);
 const renderComponent = (props = {}) => mount(
-  <Button href={href} {...props}>
+  <Button {...props}>
     {children}
   </Button>
 );
@@ -20,6 +19,11 @@ describe('<Button />', () => {
   it('should render a <button> tag to change route if the handleRoute prop is specified', () => {
     const renderedComponent = renderComponent({ handleRoute });
     expect(renderedComponent.find('button').length).toEqual(1);
+  });
+
+  it('should render a <a> tag if href prop is specified', () => {
+    const renderedComponent = renderComponent({ href: '#' });
+    expect(renderedComponent.find('a').length).toEqual(1);
   });
 
   it('should have children', () => {
