@@ -4,11 +4,12 @@ import styled from 'styled-components';
 import {
   baseColor,
   gray,
-  black,
   white,
   curtainStickyWidth,
   curtainWidth,
 } from 'variables';
+
+import { menuClose } from '../TableMenu/styles';
 
 export const CurtainWrapper = styled.div`
   width: ${curtainWidth};
@@ -20,62 +21,54 @@ export const CurtainWrapper = styled.div`
   z-index: 6;
   padding: 20px;
   transition: .15s ease left;
-  
+
   @media (min-width: ${curtainStickyWidth}) {
     left: 0px;
   }
 `;
 
-export const CurtainToggler = styled.div`
+export const CurtainTogglerWrapper = styled.div`
   position: absolute;
+  ${(props) => props.isOpen ?
+    `top: 40px;
+    left: 356px;`
+    :
+    `top: 60px;
+    left: 400px;`
+  }
   cursor: pointer;
-
-  ${(props) => props.isOpen ? `
-    left : ${curtainWidth};
-    width: 100vh;
-    top: 0;
-    height: 100vh;
-  ` : `
-    top: 40px;
-    right: 0px;
-  `}
-  
-  &:before {
-    content: '';
-    display: block;
-    position: absolute;
-    border: 35px solid transparent;
-    border-top-width: 25px;
-    border-bottom-width: 25px;
-    ${(props) => props.isOpen ? `
-      border-right-color: ${black};
-      left: -70px;
-      top: 50px;
-    ` : `
-      border-left-color: ${gray};
-    `}
-  }
-    
-  &:after {
-    position: absolute;
-    display: block;
-    white-space: nowrap;
-    font-size: large;
-    color: ${white};
-    ${(props) => props.isOpen ? `
-      content: '';
-      left: 20px;
-      top: 63px;
-    ` : `
-      content: '💬 \\00a0 \\00a0 Chat';
-      left: 40px;
-      top: 13px;
-    `}
-  }
-  
+  display: flex;
+  align-items: center;
+  color: ${gray};
   @media (min-width: ${curtainStickyWidth}) {
     display: none;
   }
+`;
+
+export const ToggleTriangle = styled.span`
+  width: 70px;
+  height: 70px;
+  clip: rect(auto, 70px, auto, 50px);
+  transform: translateX(-50px);
+  &:before {
+    content: "";
+    position: absolute;
+    top: 10px;
+    bottom: 10px;
+    left: 10px;
+    right: 10px;
+    background: ${menuClose};
+    transform: rotate(-45deg);
+  }
+`;
+
+export const ToggleIcon = styled.i`
+  margin-left: -36px;
+`;
+
+export const ToggleText = styled.span`
+  padding: 4px 0 0 7px;
+  font-size: 16px;
 `;
 
 const CurtainHeaderContainer = styled.div`
