@@ -12,9 +12,9 @@ import {
   white,
 } from '../../variables';
 
-const StyledIcon = styled.i`
+export const StyledIcon = styled.i`
   font-size: 5em !important;
-  color: ${baseColor}
+  color: ${baseColor};
   padding: 0.1em;
   &:hover {
     color: ${white};
