@@ -21,25 +21,31 @@ const Td = styled.td`
   border-top: 1px solid #eceeef;
 `;
 
-function ListItem(props) {
-  let data = (<td></td>);
-  if (props.values) {
-    data = props.values.map((val, i) => {
-      if (i === 0) {
-        return (<Th key={i}>{val}</Th>);
-      }
-      return (<Td key={i}>{val}</Td>);
-    });
-  }
+function ListItem({ values, columnsStyle = {} }) {
   return (
     <Tr>
-      {data}
+      {!values && <td />}
+      {values && values.map((val, i) => {
+        const Comp = i === 0 ? Th : Td;
+        return (
+          <Comp
+            key={i}
+            style={columnsStyle[i]}
+          >
+            {typeof val === 'number'
+              ? String(val).replace(/^-/, '−')
+              : val
+            }
+          </Comp>
+        );
+      })}
     </Tr>
   );
 }
 
 ListItem.propTypes = {
   values: React.PropTypes.array,
+  columnsStyle: React.PropTypes.object,
 };
 
 export default ListItem;
