@@ -13,11 +13,10 @@ import {
 } from 'variables';
 
 export const TableContainer = styled.div`
-  background-image: ${backgroundTable};
-  margin-top: 18px;
+  margin-top: 5%;
   margin-right: auto;
   margin-left: auto;
-  
+  height: 100%;
   @media (min-width: ${curtainStickyWidth}) {
     margin-left: ${curtainWidth};
   }
@@ -25,6 +24,7 @@ export const TableContainer = styled.div`
 
 export const TableHeader = styled.div`
    position: relative;
+   
    text-align: left;
    color: ${white};
    padding-left: 10em;
@@ -43,7 +43,7 @@ export const Wrapper = styled.div`
   }
 `;
 
-export const Winner = styled.div`
+export const Winner = styled.div` 
   left: 50%;
   top: 100%;
   position: absolute;
@@ -56,8 +56,7 @@ export const Winner = styled.div`
 
 export const TableAndChairs = styled.div`
   position: relative;
-  width: 100%
-  padding-bottom: 50%;
+  width: 100%;
 `;
 
 export const HandBox = styled.div`
@@ -73,11 +72,10 @@ export const HandBox = styled.div`
 
 export const PokerTable = styled.div`
   position: absolute;
-  margin-top: 10%;
-  border-radius: 50%;
   transform: translateX(-50%);
+  background-image: ${backgroundTable};
+  top: -20%;
   left: 50%;
-  width: 50%;
-  height: 40%;
-  z-index: 5;
+  width: 100%;
+  padding: 20%;
 `;
