@@ -1,6 +1,7 @@
 import { bindActionCreators } from 'redux';
 import { contractMethodCall, contractTxSend } from './actions';
 import { getWeb3 } from './sagas';
+import { last } from '../../utils/last';
 
 function degrade(fn, fallback) {
   try {
@@ -26,14 +27,20 @@ function generateContractInstanceApi({ abi, address, getState, dispatch }) {
     const actions = bindActionCreators({
       // dispatches action to read contract method results and write into store
       call: (...args) => contractMethodCall({
-        args, address, key: getMethodKey({ methodName, args }), method: contractInstance[methodName].call,
+        args,
+        address,
+        key: getMethodKey({ methodName, args }),
+        method: contractInstance[methodName].call,
       }),
       // creates receipt for to invoke contract through account controller
       sendTransaction: (...args) => contractTxSend({
+        args,
+        methodName,
         key: getMethodKey({ methodName, args }),
         dest: address,
         data: contractInstance[methodName].getData(...args),
         privKey: getState().get('privKey'),
+        callback: typeof last(args) === 'function' ? last(args) : undefined,
       }),
     }, dispatch);
     // base getter
