@@ -1,12 +1,11 @@
-import Web3 from 'web3';
 import ethUtil from 'ethereumjs-util';
-import { takeLatest, select, actionChannel, put, fork, take, takeEvery, call, cancelled } from 'redux-saga/effects';
+import { takeLatest, select, actionChannel, put, fork, take, takeEvery, call } from 'redux-saga/effects';
 import { delay, eventChannel, END } from 'redux-saga';
 import fetch from 'isomorphic-fetch';
 import Raven from 'raven-js';
 import { Receipt } from 'poker-helper';
+import * as storageService from '../../services/localStorage';
 
-import WebsocketProvider from '../../services/wsProvider';
 import { createBlocky } from '../../services/blockies';
 import { nickNameByAddress } from '../../services/nicknames';
 import {
@@ -16,6 +15,8 @@ import {
   ABI_ACCOUNT_FACTORY,
 } from '../../app.config';
 
+import { addEventsDate, getWeb3, isUserEvent } from './utils';
+
 import {
   WEB3_CONNECT,
   WEB3_METHOD_CALL,
@@ -23,7 +24,6 @@ import {
   CONTRACT_TX_SEND,
   SET_AUTH,
   WEB3_CONNECTED,
-  BLOCK_NOTIFY,
   ETH_TRANSFER,
   web3Error,
   web3Connected,
@@ -35,20 +35,14 @@ import {
   contractMethodError,
   contractTxSuccess,
   contractTxError,
-  contractEvent,
+  contractEvents,
   transferETHSuccess,
   transferETHError,
 } from './actions';
 
-let web3Instance;
-const confParams = conf();
+export { getWeb3 } from './utils';
 
-export function getWeb3() {
-  if (typeof web3Instance === 'undefined') {
-    web3Instance = new Web3(new WebsocketProvider(confParams.gethUrl));
-  }
-  return web3Instance;
-}
+const confParams = conf();
 
 const getPeerCount = (web3) => (
   new Promise((resolve, reject) => {
@@ -83,7 +77,7 @@ function websocketChannel() {
       // Note: when websocket first emit this connect event, it seems to be still not initialized yet.
       // and it could cause `accountLoginSaga` get called and throw an error in web3
       if (!firstConnect) {
-        emitter(web3Connected({ web3: web3Instance, isConnected: true }));
+        emitter(web3Connected({ web3: getWeb3(), isConnected: true }));
       }
 
       firstConnect = false;
@@ -125,7 +119,7 @@ function* web3ConnectSaga() {
   try {
     yield getPeerCount(getWeb3());
     yield put(web3Connected({ isConnected: true }));
-    const tokenContract = web3Instance.eth.contract(ABI_TOKEN_CONTRACT).at(confParams.ntzAddr);
+    const tokenContract = getWeb3().eth.contract(ABI_TOKEN_CONTRACT).at(confParams.ntzAddr);
     yield call(delay, 500);
     yield fork(ethEventListenerSaga, tokenContract);
   } catch (err) {
@@ -195,7 +189,7 @@ function* accountLoginSaga() {
         id: signer,
       });
       // this reads account data from the account factory
-      const res = yield getAccount(web3Instance, signer);
+      const res = yield getAccount(getWeb3(), signer);
       const proxy = res[0];
       const controller = res[1];
       const lastNonce = res[2].toNumber();
@@ -207,7 +201,7 @@ function* accountLoginSaga() {
 
       // start listen on the account controller for events
       // mostly auth errors
-      const controllerContract = web3Instance.eth.contract(ABI_CONTROLLER).at(controller);
+      const controllerContract = getWeb3().eth.contract(ABI_CONTROLLER).at(controller);
       yield fork(ethEventListenerSaga, controllerContract);
     }
   }
@@ -239,37 +233,11 @@ function sendTx(forwardReceipt) {
   });
 }
 
-function notifyBlock() {
-  return new Promise((resolve, reject) => {
-    fetch(`${confParams.txUrl}/notify`, {
-      method: 'post',
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-      },
-      body: '{}',
-    }).then((rsp) => {
-      rsp.json().then((response) => {
-        if (rsp.status >= 200 && rsp.status < 300) {
-          resolve(response);
-        } else {
-          reject({
-            status: rsp.status,
-            message: response,
-          });
-        }
-      });
-    }).catch((error) => {
-      reject(error);
-    });
-  });
-}
-
 function* contractTransactionSendSaga() {
   const txChan = yield actionChannel(CONTRACT_TX_SEND);
   while (true) { // eslint-disable-line no-constant-condition
-    const req = yield take(txChan);
-    const { dest, key, data, privKey } = req.payload;
+    const action = yield take(txChan);
+    const { dest, key, data, privKey, callback, args, methodName } = action.payload;
     const state = yield select();
     const nonce = state.get('account').get('lastNonce') + 1;
     const controller = state.get('account').get('controller');
@@ -277,10 +245,16 @@ function* contractTransactionSendSaga() {
     // send it.
     try {
       const value = yield sendTx(forwardReceipt);
-      yield put(contractTxSuccess({ address: dest, nonce, txHash: value.txHash, key }));
+      if (callback) {
+        yield call(callback, null, value.txHash);
+      }
+      yield put(contractTxSuccess({ address: dest, nonce, txHash: value.txHash, key, args, methodName }));
     } catch (err) {
-      const error = (err.message) ? err.message : err;
-      yield put(contractTxError({ address: dest, nonce, error }));
+      const error = err.message || err;
+      if (callback) {
+        yield call(callback, error);
+      }
+      yield put(contractTxError({ address: dest, nonce, error, args, methodName, action }));
     }
   }
 }
@@ -305,40 +279,55 @@ function* transferETHSaga() {
   }
 }
 
+function* updateLoggedInStatusSaga(action) { // SET_AUTH action
+  if (!action.newAuthState.loggedIn) {
+    storageService.removeItem('privKey');
+    storageService.removeItem('email');
+  } else {
+    Raven.setUserContext({
+      email: action.newAuthState.email,
+    });
+    storageService.setItem('privKey', action.newAuthState.privKey);
+    storageService.setItem('email', action.newAuthState.email);
+  }
+}
 
 const ethEvent = (contract) => eventChannel((emitter) => {
-  const contractEvents = contract.allEvents({ fromBlock: 'latest' });
-  contractEvents.watch((error, results) => {
+  const events = contract.allEvents({ fromBlock: 'latest' });
+  events.watch((error, results) => {
     if (error) {
       emitter(END);
-      contractEvents.stopWatching();
+      events.stopWatching();
       return;
     }
     emitter(results);
   });
   return () => {
-    contractEvents.stopWatching();
+    events.stopWatching();
   };
 });
 
 export function* ethEventListenerSaga(contract) {
   const chan = yield call(ethEvent, contract);
-  try {
-    const event = yield take(chan);
-    yield put(contractEvent({ event }));
-  } finally {
-    if (yield cancelled()) {
-      chan.close();
-    }
+  while (true) { // eslint-disable-line no-constant-condition
+    try {
+      const event = yield take(chan);
+      const state = yield select();
+      const proxy = state.getIn(['account', 'proxy']);
+      if (isUserEvent(proxy)(event)) {
+        const events = yield call(addEventsDate, [event]);
+        yield put(contractEvents(events, proxy));
+      }
+    } catch (e) {} // eslint-disable-line no-empty
   }
 }
 
 // The root saga is what is sent to Redux's middleware.
 export function* accountSaga() {
   yield takeLatest(WEB3_CONNECT, web3ConnectSaga);
-  yield takeLatest(BLOCK_NOTIFY, notifyBlock);
   yield takeEvery(WEB3_METHOD_CALL, web3MethodCallSaga);
   yield takeEvery(CONTRACT_METHOD_CALL, contractMethodCallSaga);
+  yield takeEvery(SET_AUTH, updateLoggedInStatusSaga);
   yield fork(websocketSaga);
   yield fork(transferETHSaga);
   yield fork(accountLoginSaga);
