@@ -2,16 +2,16 @@
  * Created by jzobro on 20170606
  */
 // const
-export const ACTIONBAR_SET_ACTIVE = 'acebusters/ActionBar/SET_ACTIVE';
+export const ACTIONBAR_SET_TURN_COMPLETE = 'acebusters/ActionBar/SET_TURN_COMPLETE';
 export const ACTIONBAR_SET_MODE = 'acebusters/ActionBar/SET_MODE';
-export const ACTIONBAR_TOGGLE_VISIBLE = 'acebusters/ActionBar/TOGGLE_VISIBLE';
 export const ACTIONBAR_SET_BET_SLIDER = 'acebusters/ActionBar/SET_BET_SLIDER';
+export const ACTIONBAR_SET_BUTTON_ACTIVE = 'acebusters/ActionBar/SET_BUTTON_ACTIVE';
 
-// actions
-export function setActionBarActive(active) {
+// actionBar
+export function setActionBarTurnComplete(complete) {
   return {
-    type: ACTIONBAR_SET_ACTIVE,
-    active,
+    type: ACTIONBAR_SET_TURN_COMPLETE,
+    complete,
   };
 }
 
@@ -22,13 +22,17 @@ export function setActionBarMode(mode) {
   };
 }
 
+export function setActionBarButtonActive(whichBtn) {
+  return {
+    type: ACTIONBAR_SET_BUTTON_ACTIVE,
+    whichBtn,
+  };
+}
+
+// slider
 export function setActionBarBetSlider(sliderOpen) {
   return {
     type: ACTIONBAR_SET_BET_SLIDER,
     sliderOpen,
   };
-}
-
-export function toggleActionBarVisible() {
-  return { type: ACTIONBAR_TOGGLE_VISIBLE };
 }
