@@ -10,50 +10,53 @@ import makeSelectAccountData, { makeSignerAddrSelector, makeSelectPrivKey } from
 import messages from './messages';
 import { modalAdd, modalDismiss } from '../App/actions';
 import web3Connect from '../AccountProvider/web3Connect';
-import { contractEvent, accountLoaded, transferETH } from '../AccountProvider/actions';
+import { contractEvents, accountLoaded, transferETH, proxyEvents } from '../AccountProvider/actions';
+import { addEventsDate, isUserEvent } from '../AccountProvider/utils';
 import { createBlocky } from '../../services/blockies';
-import { ABI_TOKEN_CONTRACT, ABI_ACCOUNT_FACTORY, conf } from '../../app.config';
+import { ABI_TOKEN_CONTRACT, ABI_ACCOUNT_FACTORY, ABI_PROXY, ABI_TABLE_FACTORY, conf } from '../../app.config';
+import { ETH_DECIMALS, NTZ_DECIMALS, formatEth, formatNtz } from '../../utils/amountFormater';
 
 import List from '../../components/List';
+import H2 from '../../components/H2';
+import Alert from '../../components/Alert';
 import TransferDialog from '../TransferDialog';
+import ExchangeDialog from '../ExchangeDialog';
 import Container from '../../components/Container';
 import Button from '../../components/Button';
+import SubmitButton from '../../components/SubmitButton';
 import Blocky from '../../components/Blocky';
-// import FormGroup from '../../components/Form/FormGroup';
 import WithLoading from '../../components/WithLoading';
 
 import { Section } from './styles';
+import { createDashboardTxsSelector } from './selectors';
+import { txnsToList } from './txnsToList';
 
 const confParams = conf();
-// ether units: https://github.com/ethereum/web3.js/blob/0.15.0/lib/utils/utils.js#L40
-const ethDecimals = new BigNumber(10).pow(18);
-// acebuster units:
-// 1 x 10^12 - Nutz   (NTZ)
-// 1 x 10^9 - Jonyz
-// 1 x 10^6 - Helcz
-// 1 x 10^3 - Pascalz
-// 1 x 10^0 - Babz
-const ntzDecimals = new BigNumber(10).pow(12);
+
+const LOOK_BEHIND_PERIOD = 4 * 60 * 24;
 
 export class Dashboard extends React.Component { // eslint-disable-line react/prefer-stateless-function
 
   constructor(props) {
     super(props);
     this.handleNTZTransfer = this.handleNTZTransfer.bind(this);
+    this.handleNTZPurchase = this.handleNTZPurchase.bind(this);
+    this.handleNTZSell = this.handleNTZSell.bind(this);
     this.handleETHTransfer = this.handleETHTransfer.bind(this);
     this.web3 = props.web3Redux.web3;
+
     this.token = this.web3.eth.contract(ABI_TOKEN_CONTRACT).at(confParams.ntzAddr);
-    this.web3.eth.getBlockNumber((err, blockNumber) => {
-      const events = this.token.allEvents({ fromBlock: blockNumber - (4 * 60 * 24), toBlock: 'latest' });
-      events.get((error, eventList) => {
-        eventList.forEach((event) => {
-          props.contractEvent(event);
-        });
-      });
-    });
+    this.tableFactory = this.web3.eth.contract(ABI_TABLE_FACTORY).at(confParams.tableFactory);
+
+    this.tableFactory.getTables.call();
 
     if (this.props.account.proxy) {
+      this.token.floor.call();
+      this.token.ceiling.call();
+      this.token.balanceOf.call(this.props.account.proxy);
       this.web3.eth.getBalance(this.props.account.proxy);
+      this.watchProxyEvents(this.props.account.proxy);
+      this.watchTokenEvents(this.props.account.proxy);
     }
   }
 
@@ -64,13 +67,32 @@ export class Dashboard extends React.Component { // eslint-disable-line react/pr
   }
 
   componentWillReceiveProps(nextProps) {
-    const balance = this.token.balanceOf(this.props.account.proxy);
-    if (!balance && nextProps.account.proxy) {
+    if (this.props.account.proxy === undefined && nextProps.account.proxy) {
+      this.token.floor.call();
+      this.token.ceiling.call();
+      this.watchProxyEvents(nextProps.account.proxy);
+      this.watchTokenEvents(nextProps.account.proxy);
       this.token.balanceOf.call(nextProps.account.proxy);
+      this.web3.eth.getBalance(nextProps.account.proxy);
     }
 
-    if (this.props.account.proxy === undefined && nextProps.account.proxy) {
-      this.web3.eth.getBalance(nextProps.account.proxy);
+    if (this.props.dashboardTxs.txError !== nextProps.dashboardTxs.txError && nextProps.dashboardTxs.txError) {
+      this.props.modalAdd(
+        <div>
+          <H2>
+            <FormattedMessage {...messages.transactionErrorTitle} />
+          </H2>
+          <p>{nextProps.dashboardTxs.txError}</p>
+          <SubmitButton
+            onClick={() => {
+              this.props.dispatch(nextProps.dashboardTxs.failedTxAction);
+              this.props.modalDismiss();
+            }}
+          >
+            <FormattedMessage {...messages.retryTransaction} />
+          </SubmitButton>
+        </div>
+      );
     }
 
     // Note: listen to AccountFactory's AccountCreated Event if proxy address is not ready
@@ -81,20 +103,75 @@ export class Dashboard extends React.Component { // eslint-disable-line react/pr
     }
   }
 
-  handleNTZTransfer(to, amount) {
-    this.token.transfer.sendTransaction(
-      to,
-      `0x${new BigNumber(amount).mul(ntzDecimals).toString(16)}`
-    );
-    this.props.modalDismiss();
+  watchProxyEvents(proxyAddr) {
+    const web3 = getWeb3();
+    this.proxy = web3.eth.contract(ABI_PROXY).at(proxyAddr);
+
+    this.web3.eth.getBlockNumber((err, blockNumber) => {
+      this.proxy.allEvents({
+        fromBlock: blockNumber - LOOK_BEHIND_PERIOD,
+        toBlock: 'latest',
+      }).get((error, eventList) => {
+        addEventsDate(eventList.filter(isUserEvent(proxyAddr)))
+          .then((events) => this.props.proxyEvents(events, proxyAddr));
+      });
+    });
+
+    this.proxy.allEvents({
+      toBlock: 'latest',
+    }).watch((error, event) => {
+      if (!error && event) {
+        addEventsDate([event])
+          .then((events) => this.props.proxyEvents(events, proxyAddr));
+        this.web3.eth.getBalance(proxyAddr);
+      }
+    });
   }
 
-  handleETHTransfer(dest, amount) {
-    this.props.transferETH({
-      dest,
-      amount: `0x${new BigNumber(amount).mul(ethDecimals).toString(16)}`,
+  watchTokenEvents(proxyAddr) {
+    this.web3.eth.getBlockNumber((err, blockNumber) => {
+      this.token.allEvents({
+        fromBlock: blockNumber - LOOK_BEHIND_PERIOD,
+        toBlock: 'latest',
+      }).get((error, eventList) => {
+        addEventsDate(eventList.filter(isUserEvent(proxyAddr)))
+          .then((events) => this.props.contractEvents(events, proxyAddr));
+      });
     });
-    this.props.modalDismiss();
+
+    this.token.allEvents({
+      toBlock: 'latest',
+    }).watch((watchError, event) => {
+      if (!watchError && isUserEvent(proxyAddr)(event)) {
+        this.token.balanceOf.call(proxyAddr);
+        this.web3.eth.getBalance(proxyAddr);
+        const { pendingSell = [] } = this.props.dashboardTxs;
+
+        if (pendingSell.indexOf(event.transactionHash) > -1) {
+          this.token.transferFrom.sendTransaction(
+            confParams.ntzAddr,
+            proxyAddr,
+            0,
+            { from: proxyAddr }
+          );
+        }
+      }
+    });
+
+    // Check if we have unfinished sell
+    this.token.allowance.callPromise(
+      confParams.ntzAddr,
+      proxyAddr,
+    ).then((value) => {
+      if (!value.eq(0)) {
+        this.token.transferFrom.sendTransaction(
+          confParams.ntzAddr,
+          proxyAddr,
+          0,
+          { from: proxyAddr }
+        );
+      }
+    });
   }
 
   watchAccountCreated() {
@@ -118,19 +195,52 @@ export class Dashboard extends React.Component { // eslint-disable-line react/pr
     });
   }
 
+  handleNTZTransfer(to, amount) {
+    this.token.transfer.sendTransaction(
+      to,
+      new BigNumber(amount).mul(NTZ_DECIMALS)
+    );
+    this.props.modalDismiss();
+  }
+
+  handleNTZPurchase(amount) {
+    this.props.transferETH({
+      dest: confParams.ntzAddr,
+      amount: new BigNumber(amount).mul(ETH_DECIMALS),
+    });
+    this.props.modalDismiss();
+  }
+
+  handleNTZSell(amount) {
+    this.token.transfer.sendTransaction(
+      confParams.ntzAddr,
+      new BigNumber(amount).mul(NTZ_DECIMALS),
+      { from: this.props.account.proxy }
+    );
+    this.props.modalDismiss();
+  }
+
+  handleETHTransfer(dest, amount) {
+    this.props.transferETH({
+      dest,
+      amount: new BigNumber(amount).mul(ETH_DECIMALS),
+    });
+    this.props.modalDismiss();
+  }
+
   render() {
     const qrUrl = `ether:${this.props.account.proxy}`;
     const weiBalance = this.web3.eth.balance(this.props.account.proxy);
-    const ethBalance = (weiBalance) ? weiBalance.div(ethDecimals) : weiBalance;
-    const babBalance = this.token.balanceOf(this.props.account.proxy);
-    const ntzBalance = (babBalance) ? babBalance.div(ntzDecimals) : babBalance;
+    const floor = this.token.floor();
+    const ceiling = this.token.ceiling();
+    const babzBalance = this.token.balanceOf(this.props.account.proxy);
+    const tables = this.tableFactory.getTables();
 
-    const listPending = pendingToList(this.props.account.pending);
-
-    let listTxns = null;
-    if (this.props.account[confParams.ntzAddr]) {
-      listTxns = txnsToList(this.props.account[confParams.ntzAddr].transactions, this.props.account.proxy);
-    }
+    const listTxns = txnsToList(
+      this.props.dashboardTxs.dashboardEvents,
+      tables,
+      this.props.account.proxy
+    );
 
     return (
       <Container>
@@ -138,7 +248,7 @@ export class Dashboard extends React.Component { // eslint-disable-line react/pr
 
         <Section>
           <Blocky blocky={createBlocky(this.props.signerAddr)} />
-          <h3> Your address:</h3>
+          <h3>Your address:</h3>
 
           <WithLoading
             isLoading={!this.props.account.proxy || this.props.account.proxy === '0x'}
@@ -147,7 +257,12 @@ export class Dashboard extends React.Component { // eslint-disable-line react/pr
           >
             <p> { this.props.account.proxy } </p>
             <QRCode value={qrUrl} size={120} />
+
+            <Alert theme="danger">
+              <FormattedMessage {...messages.ethAlert} />
+            </Alert>
           </WithLoading>
+
         </Section>
 
         <Section>
@@ -155,29 +270,72 @@ export class Dashboard extends React.Component { // eslint-disable-line react/pr
           <p>
             <span>Balance: </span>
             <WithLoading
-              isLoading={ntzBalance === undefined || ntzBalance === null}
+              isLoading={!babzBalance}
               loadingSize="14px"
               type="inline"
               styles={{ layout: { marginLeft: '15px' } }}
             >
-              <span>{ntzBalance && ntzBalance.toString()} NTZ</span>
+              <span>{babzBalance && formatNtz(babzBalance)} NTZ</span>
             </WithLoading>
           </p>
-          <Button
-            align="left"
-            onClick={() => {
-              this.props.modalAdd(
-                <TransferDialog
-                  handleTransfer={this.handleNTZTransfer}
-                  amountUnit="NTZ"
-                />
-              );
-            }}
-            size="medium"
-            icon="fa fa-money"
-          >
-            TRANSFER
-          </Button>
+          {babzBalance &&
+            <Button
+              align="left"
+              onClick={() => {
+                this.props.modalAdd(
+                  <TransferDialog
+                    handleTransfer={this.handleNTZTransfer}
+                    maxAmount={babzBalance.div(NTZ_DECIMALS)}
+                    amountUnit="NTZ"
+                  />
+                );
+              }}
+              size="medium"
+              icon="fa fa-money"
+            >
+              TRANSFER
+            </Button>
+          }
+          {babzBalance && floor &&
+            <Button
+              align="left"
+              onClick={() => {
+                this.props.modalAdd(
+                  <ExchangeDialog
+                    title={<FormattedMessage {...messages.sellTitle} />}
+                    amountUnit="ntz"
+                    calcExpectedAmount={(amount) => new BigNumber(amount).div(floor)}
+                    handleExchange={this.handleNTZSell}
+                    maxAmount={babzBalance.div(NTZ_DECIMALS)}
+                  />
+                );
+              }}
+              size="medium"
+              icon="fa fa-money"
+            >
+              SELL
+            </Button>
+          }
+          {weiBalance && ceiling &&
+            <Button
+              align="left"
+              onClick={() => {
+                this.props.modalAdd(
+                  <ExchangeDialog
+                    title={<FormattedMessage {...messages.purchaseTitle} />}
+                    amountUnit="eth"
+                    calcExpectedAmount={(amount) => ceiling.mul(amount)}
+                    handleExchange={this.handleNTZPurchase}
+                    maxAmount={weiBalance.div(ETH_DECIMALS)}
+                  />
+                );
+              }}
+              size="medium"
+              icon="fa fa-money"
+            >
+              PURCHASE
+            </Button>
+          }
         </Section>
 
         <Section>
@@ -185,96 +343,91 @@ export class Dashboard extends React.Component { // eslint-disable-line react/pr
           <p>
             <span>Balance: </span>
             <WithLoading
-              isLoading={ethBalance === undefined}
+              isLoading={!weiBalance}
               loadingSize="14px"
               type="inline"
               styles={{ layout: { marginLeft: '15px' } }}
             >
-              <span>{ethBalance && ethBalance.toString()} ETH</span>
+              <span>{weiBalance && formatEth(weiBalance)} ETH</span>
             </WithLoading>
           </p>
-          <Button
-            align="left"
-            onClick={() => {
-              this.props.modalAdd(
-                <TransferDialog
-                  handleTransfer={this.handleETHTransfer}
-                  amountUnit="ETH"
-                />
-              );
-            }}
-            size="medium"
-            icon="fa fa-money"
-          >
-            TRANSFER
-          </Button>
+          {weiBalance &&
+            <Button
+              align="left"
+              onClick={() => {
+                this.props.modalAdd(
+                  <TransferDialog
+                    handleTransfer={this.handleETHTransfer}
+                    maxAmount={weiBalance.div(ETH_DECIMALS)}
+                    amountUnit="ETH"
+                  />
+                );
+              }}
+              size="medium"
+              icon="fa fa-money"
+            >
+              TRANSFER
+            </Button>
+          }
         </Section>
 
         <Section>
-          <h2><FormattedMessage {...messages.pending} /></h2>
-          <List items={listPending} headers={['#', 'txHash']} noDataMsg="No Pending Transactions" />
           <h2><FormattedMessage {...messages.included} /></h2>
-          <List items={listTxns} headers={['txHash', 'from', 'to', 'amount']} noDataMsg="No Transactions Yet" />
+          <List
+            items={listTxns}
+            headers={[
+              '',
+              'Address',
+              'Date',
+              '',
+              'Amount',
+              '',
+            ]}
+            columnsStyle={{
+              0: { width: 20 },
+              1: { textAlign: 'left', width: 10, whiteSpace: 'nowrap' },
+              2: { width: 20 },
+              3: { textAlign: 'left', whiteSpace: 'nowrap' },
+              4: { textAlign: 'right', whiteSpace: 'nowrap' },
+              5: { width: '100%', textAlign: 'left' },
+            }}
+            noDataMsg="No Transactions Yet"
+          />
         </Section>
-
       </Container>
     );
   }
 }
 
-const pendingToList = (pending) => {
-  let list = [];
-  if (pending) {
-    list = Object.keys(pending).map((key) => [key, pending[key].txHash]);
-  }
-  return list;
-};
-
-const txnsToList = (txns, proxyAddr) => {
-  if (txns) {
-    return Object.keys(txns)
-      .filter((key) => (
-        (key && txns[key] && txns[key].from && txns[key].to) &&
-        (txns[key].from === proxyAddr || txns[key].to === proxyAddr)
-      ))
-      .map((key) => ({
-        txHash: key,
-        blockNumber: txns[key].blockNumber,
-        from: txns[key].from,
-        to: txns[key].to,
-        value: (txns[key].to === proxyAddr) ? txns[key].value : txns[key].value * -1,
-      }))
-      .map((entry) => [entry.txHash.substring(2, 8), entry.from.substring(2, 8), entry.to.substring(2, 8), entry.value]);
-  }
-
-  return null;
-};
-
 Dashboard.propTypes = {
   modalAdd: PropTypes.func,
   transferETH: PropTypes.func,
+  proxyEvents: PropTypes.func,
   modalDismiss: PropTypes.func,
-  contractEvent: PropTypes.func,
+  contractEvents: PropTypes.func,
   accountLoaded: PropTypes.func,
   web3Redux: PropTypes.any,
   signerAddr: PropTypes.string,
-  account: PropTypes.any,
+  account: PropTypes.object,
+  dashboardTxs: PropTypes.object,
   privKey: PropTypes.string,
+  dispatch: PropTypes.func,
 };
 
 const mapStateToProps = createStructuredSelector({
   account: makeSelectAccountData(),
+  dashboardTxs: createDashboardTxsSelector(),
   signerAddr: makeSignerAddrSelector(),
   privKey: makeSelectPrivKey(),
 });
-
 
 function mapDispatchToProps() {
   return {
     modalAdd,
     modalDismiss,
     transferETH,
-    contractEvent: (event) => contractEvent({ event }),
+    proxyEvents,
+    contractEvents,
     accountLoaded,
   };
 }
