@@ -10,12 +10,68 @@ export default defineMessages({
     id: 'app.containers.Dashboard.header',
     defaultMessage: 'Account Dashboard',
   },
-  pending: {
-    id: 'app.containers.Dashboard.pending',
-    defaultMessage: 'Pending Transactions:',
-  },
   included: {
     id: 'app.containers.Dashboard.included',
     defaultMessage: 'Transaction History:',
+  },
+  transactionErrorTitle: {
+    id: 'app.containers.Dashboard.transactionErrorTitle',
+    defaultMessage: 'Transaction error',
+  },
+  retryTransaction: {
+    id: 'app.containers.Dashboard.retryTransaction',
+    defaultMessage: 'Retry',
+  },
+  ethAlert: {
+    id: 'app.containers.Dashboard.ethAlert',
+    defaultMessage: 'Never send real ether to this address. It is a Rinkeby Testnet Address. Only send Rinkaby TestEther to this Address. Otherwise your Ether will be lost',
+  },
+  sellTitle: {
+    id: 'app.containers.Dashboard.sellTitle',
+    defaultMessage: 'Sell NTZ',
+  },
+  purchaseTitle: {
+    id: 'app.containers.Dashboard.purchaseTitle',
+    defaultMessage: 'Purchase NTZ',
+  },
+  tableJoin: {
+    id: 'app.containers.Dashboard.tableJoin',
+    defaultMessage: 'Table join',
+  },
+  tableLeave: {
+    id: 'app.containers.Dashboard.tableLeave',
+    defaultMessage: 'Table leave',
+  },
+  sellStart: {
+    id: 'app.containers.Dashboard.sellStart',
+    defaultMessage: 'Sell start',
+  },
+  sellEnd: {
+    id: 'app.containers.Dashboard.sellEnd',
+    defaultMessage: 'Sell end',
+  },
+  purchaseStart: {
+    id: 'app.containers.Dashboard.purchaseStart',
+    defaultMessage: 'Purchase start',
+  },
+  purchaseEnd: {
+    id: 'app.containers.Dashboard.purchaseEnd',
+    defaultMessage: 'Purchase end',
+  },
+  nutzContract: {
+    id: 'app.containers.Dashboard.nutzContract',
+    defaultMessage: 'Nutz contract',
+  },
+  me: {
+    id: 'app.containers.Dashboard.me',
+    defaultMessage: 'Nutz contract',
+  },
+  tableAddress: {
+    id: 'app.containers.Dashboard.tableAddress',
+    defaultMessage: 'Table {address}',
+  },
+  trasnferStatus: {
+    id: 'app.containers.Dashboard.trasnferStatus',
+    defaultMessage: 'Transfer',
   },
 });
