@@ -3,13 +3,12 @@ import { connect } from 'react-redux';
 import Slider from 'react-rangeslider';
 import 'react-rangeslider/lib/index.css';
 import { createStructuredSelector } from 'reselect';
-import Button from '../../components/Button';
+import SubmitButton from '../../components/SubmitButton';
 import H2 from '../../components/H2';
 
 import { makeSbSelector } from '../Table/selectors';
-import {
-  makeSelectProxyAddr,
-} from '../AccountProvider/selectors';
+import { makeSelectProxyAddr } from '../AccountProvider/selectors';
+import { formatNtz } from '../../utils/amountFormater';
 
 export class JoinDialog extends React.Component { // eslint-disable-line react/prefer-stateless-function
 
@@ -34,13 +33,13 @@ export class JoinDialog extends React.Component { // eslint-disable-line react/p
   render() {
     const min = this.props.sb * 40;
     const tableMax = this.props.sb * 200;
-    const max = (this.props.balance < tableMax) ? this.props.balance : tableMax;
+    const max = (this.props.balance < tableMax) ? this.props.balance - (this.props.balance % this.props.sb) : tableMax;
     if (this.props.balance < min) {
       return (
         <div style={{ minWidth: '20em' }}>
           <H2>Sorry!</H2>
           <p>Your balance is not sufficient to join this table!</p>
-          <Button onClick={this.props.modalDismiss}>OK</Button>
+          <SubmitButton onClick={this.props.modalDismiss}>OK</SubmitButton>
         </div>
       );
     }
@@ -52,13 +51,13 @@ export class JoinDialog extends React.Component { // eslint-disable-line react/p
           tooltip={false}
           min={min}
           max={max}
-          step={1}
+          step={this.props.sb}
           onChange={this.updateAmount}
         >
         </Slider>
-        <div> Max: {max}</div>
-        <div>{ (this.state) ? this.state.amount : min }</div>
-        <Button onClick={this.handleSubmit}>Join</Button>
+        <div>Max: {formatNtz(max)} NTZ</div>
+        <div>{ (this.state) ? formatNtz(this.state.amount) : formatNtz(min) } NTZ</div>
+        <SubmitButton onClick={this.handleSubmit}>Join</SubmitButton>
       </div>
     );
   }
