@@ -2,15 +2,18 @@ import React from 'react';
 import { connect } from 'react-redux';
 import ReCAPTCHA from 'react-google-recaptcha';
 import { Form, Field, reduxForm, propTypes } from 'redux-form/immutable';
+import { FormattedMessage } from 'react-intl';
 
 // components
 import Container from '../../components/Container';
 import FormGroup from '../../components/Form/FormGroup';
+import FormField from '../../components/Form/FormField';
 import Label from '../../components/Label';
-import Input from '../../components/Input';
+import { CheckBox } from '../../components/Input';
 import Button from '../../components/Button';
 import H1 from '../../components/H1';
 import { ErrorMessage, WarningMessage } from '../../components/FormMessages';
+import messages from './messages';
 
 import { register } from './actions';
 
@@ -37,6 +40,10 @@ const validate = (values) => {
     }
   }
 
+  if (!values.get('terms')) {
+    errors.captchaResponse = 'Required';
+  }
+
   return errors;
 };
 
@@ -56,10 +63,10 @@ const Captcha = (props) => (
   </div>
 );
 
-const renderField = ({ input, label, type, meta: { touched, error, warning } }) => (
+const renderCheckBox = ({ input, label, type, meta: { touched, error, warning } }) => (
   <FormGroup>
-    <Label htmlFor={input.name}>{label}</Label>
-    <Input {...input} placeholder={label} type={type} />
+    <CheckBox {...input} placeholder={label} type={type} />
+    <Label htmlFor={input.name}><FormattedMessage {...messages.terms.agree} /> {label}</Label>
     {touched && ((error && <ErrorMessage error={error} />) || (warning && <WarningMessage warning={warning} />))}
   </FormGroup>
 );
@@ -83,7 +90,7 @@ export class RegisterPage extends React.Component { // eslint-disable-line react
 
   render() {
     const { error, invalid, submitting, handleSubmit, asyncValidating } = this.props;
-
+    const termsLink = (<a href="http://www.acebusters.com/terms_of_use.html" target="_blank"><FormattedMessage {...messages.terms} /></a>);
     return (
       <Container>
         <div>
@@ -91,15 +98,21 @@ export class RegisterPage extends React.Component { // eslint-disable-line react
           <Form
             onSubmit={handleSubmit(this.handleSubmit)}
           >
-            <Field name="email" type="text" component={renderField} label="e-mail" />
+            <Field name="email" type="text" component={FormField} label="e-mail" />
             <Field
               name="referral"
               type="text"
-              component={renderField}
+              component={FormField}
               label="referral code"
             />
             <Field name="captchaResponse" component={Captcha} />
             {error && <ErrorMessage error={error} />}
+            <Field
+              name="terms"
+              type="checkbox"
+              component={renderCheckBox}
+              label={termsLink}
+            />
             <Button type="submit" disabled={submitting || invalid || asyncValidating} size="large">
               { (!submitting) ? 'Register' : 'Please wait ...' }
             </Button>
@@ -123,7 +136,11 @@ function mapDispatchToProps(dispatch) {
 }
 
 // Which props do we want to inject, given the global state?
-const mapStateToProps = () => ({});
+const mapStateToProps = (state, props) => ({
+  initialValues: {
+    referral: props.params.refCode,
+  },
+});
 
 // Wrap the component to inject dispatch and state into it
 export default connect(mapStateToProps, mapDispatchToProps)(
