@@ -135,12 +135,12 @@ const makeFoldedSelector = () => createSelector(
 
 const makeCoordsSelector = () => createSelector(
   [makeLineupSelector(), posSelector],
-  (lineup, pos) => (lineup && pos > -1) ? SEAT_COORDS[lineup.toJS().length.toString()][pos] : null
+  (lineup, pos) => (lineup && pos > -1) ? SEAT_COORDS[pos] : null
 );
 
 const makeAmountCoordsSelector = () => createSelector(
   [makeLineupSelector(), posSelector],
-  (lineup, pos) => (lineup && pos > -1) ? AMOUNT_COORDS[lineup.toJS().length.toString()][pos] : null
+  (lineup, pos) => (lineup && pos > -1) ? AMOUNT_COORDS[pos] : null
 );
 
 const makeBlockySelector = () => createSelector(
