@@ -18,6 +18,7 @@ import Button from '../../components/Button';
 import Slides from '../../components/Slides';
 import { nickNameByAddress } from '../../services/nicknames';
 import messages from './messages';
+import { formatNtz } from '../../utils/amountFormater';
 
 // config data
 import {
@@ -49,8 +50,6 @@ import makeSelectAccountData, {
 import {
   makeLastReceiptSelector,
 } from '../Seat/selectors';
-
-import { blockNotify } from '../AccountProvider/actions';
 
 import {
   makeTableDataSelector,
@@ -118,6 +117,7 @@ export class Table extends React.PureComponent { // eslint-disable-line react/pr
       tableAddr: this.tableAddr,
       handId,
     });
+    this.tableService = new TableService(this.props.params.tableAddr, this.props.privKey);
   }
 
   componentWillReceiveProps(nextProps) {
@@ -136,8 +136,7 @@ export class Table extends React.PureComponent { // eslint-disable-line react/pr
 
       if (timeOut > 0) {
         this.timeOut = setTimeout(() => {
-          const table = new TableService(this.props.params.tableAddr);
-          table.timeOut().then((res) => {
+          this.tableService.timeOut().then((res) => {
             Raven.captureMessage(`timeout: ${res}`, { tags: {
               tableAddr: this.props.params.tableAddr,
               handId,
@@ -314,7 +313,6 @@ export class Table extends React.PureComponent { // eslint-disable-line react/pr
     const handId = parseInt(this.props.params.handId, 10);
     const state = this.props.state;
     const exitHand = (state !== 'waiting') ? handId : handId - 1;
-    const table = new TableService(this.props.params.tableAddr, this.props.privKey);
     this.props.setExitHand(this.tableAddr, this.props.params.handId, pos, exitHand);
     const statusElement = (<div>
       <p>
@@ -327,7 +325,7 @@ export class Table extends React.PureComponent { // eslint-disable-line react/pr
     this.props.modalDismiss();
     this.props.modalAdd(statusElement);
 
-    return table.leave(exitHand).catch((err) => {
+    return this.tableService.leave(exitHand).catch((err) => {
       Raven.captureException(err, { tags: {
         tableAddr: this.props.params.tableAddr,
         handId,
@@ -364,16 +362,8 @@ export class Table extends React.PureComponent { // eslint-disable-line react/pr
         ];
 
         if (result.args && result.args.addr === this.props.proxyAddr) {
-          // notify backend about new block
-          this.props.blockNotify();
-          // show modal
-
-          const statusElement = (<div>
-            <h2>Join Successful!</h2>
-            <Button onClick={this.handleJoinComplete}>OK!</Button>
-          </div>);
-          this.props.modalDismiss();
-          this.props.modalAdd(statusElement);
+          // notify backend about change in lineup
+          this.tableService.lineup();
         }
 
         // update lineup when join successful
@@ -381,6 +371,12 @@ export class Table extends React.PureComponent { // eslint-disable-line react/pr
           this.props.lineupReceived(...lineupReceivedArgs(rsp));
         });
 
+        break;
+      }
+
+      case 'Leave': {
+        // notify backend about change in lineup
+        this.tableService.lineup();
         break;
       }
 
@@ -475,7 +471,7 @@ export class Table extends React.PureComponent { // eslint-disable-line react/pr
     if (this.props.winners && this.props.winners.length > 0) {
       winners = this.props.winners.map((winner, index) => {
         const handString = (winner.hand) ? `with ${winner.hand}` : '';
-        return (<div key={index}>{nickNameByAddress(winner.addr)} won {winner.amount} {handString}</div>);
+        return (<div key={index}>{nickNameByAddress(winner.addr)} won {formatNtz(winner.amount)}  NTZ {handString}</div>);
       });
     }
     const sb = (this.props.data && this.props.data.get('smallBlind')) ? this.props.data.get('smallBlind') : 0;
@@ -493,6 +489,7 @@ export class Table extends React.PureComponent { // eslint-disable-line react/pr
           sitout={this.props.sitout}
           board={board}
           seats={seats}
+          hand={this.props.hand}
           onLeave={() => this.handleLeave(this.props.myPos)}
           onSitout={this.handleSitout}
         >
@@ -513,7 +510,6 @@ export function mapDispatchToProps() {
     setExitHand: (tableAddr, handId, pos, exitHand) => (setExitHand(tableAddr, handId, pos, exitHand)),
     updateReceived: (tableAddr, hand) => (updateReceived(tableAddr, hand)),
     addMessage: (message, tableAddr, privKey, created) => (addMessage(message, tableAddr, privKey, created)),
-    blockNotify: () => (blockNotify()),
   };
 }
 
@@ -549,6 +545,8 @@ Table.propTypes = {
   params: React.PropTypes.object,
   privKey: React.PropTypes.string,
   lastReceipt: React.PropTypes.string,
+  latestHand: React.PropTypes.any,
+  missingHands: React.PropTypes.any,
   sitoutAmount: React.PropTypes.number,
   proxyAddr: React.PropTypes.string,
   signerAddr: React.PropTypes.string,
@@ -556,7 +554,6 @@ Table.propTypes = {
   data: React.PropTypes.any,
   myPos: React.PropTypes.any,
   modalAdd: React.PropTypes.func,
-  blockNotify: React.PropTypes.func,
   handRequest: React.PropTypes.func,
   pendingToggle: React.PropTypes.func,
   setExitHand: React.PropTypes.func,
