@@ -14,7 +14,7 @@ const dllPlugin = pkg.dllPlugin;
 
 const plugins = [
   new webpack.HotModuleReplacementPlugin(), // Tell webpack we want hot reloading
-  new webpack.NoErrorsPlugin(),
+  new webpack.NoEmitOnErrorsPlugin(),
   new HtmlWebpackPlugin({
     inject: true, // Inject all files that are generated by webpack, e.g. bundle.js
     templateContent: templateContent(), // eslint-disable-line no-use-before-define
@@ -52,7 +52,8 @@ module.exports = require('./webpack.base.babel')({
   },
 
   // Emit a source map for easier debugging
-  devtool: 'inline-source-map',
+  // See https://webpack.js.org/configuration/devtool/#devtool
+  devtool: 'eval-source-map',
 
   performance: {
     hints: false,
